requirements.txt
Django == 5.1.4
django-markdownx == 4.0.7
djangorestframework == 3.15.2
django-filter == 24.3
django-cors-headers == 4.7.0
djangorestframework-simplejwt[crypto] == 5.4.0
pyotp == 2.9.0
qrcode[pil] == 7.4.2
